Native Windows multi-touch support and fixes

+++ Core/Contents/Source/PolySound.cpp
@@ -312,7 +312,7 @@ ALuint Sound::loadOGG(const String& fileName) {
 	} while (bytes > 0);
 	ov_clear(&oggFile);
 	
-	sampleLength = buffer.size() / sizeof(uint16_t);
+	sampleLength = buffer.size() / sizeof(unsigned int);
 	
 	alBufferData(bufferID, format, &buffer[0], static_cast<ALsizei>(buffer.size()), freq);
